Postpone initial API call until 30 seconds after startup
On mobile devices, Obsidian takes a few seconds to start up. We do not
want to add to this load with a call to the Gitlab API.

Therefore, increase the startup delay from 5 seconds to 30 seconds, so
that it can be handled once all the application start jobs have
finished.